cleaned up some testing
