+++ src/asgard_basis.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "asgard_block_matrix.hpp"
-#include "asgard_pde.hpp"
+#include "asgard_pde_utils.hpp"
 #include "asgard_wavelet_basis.hpp"
 
 namespace asgard
