+++ src/asgard_solver_tests.cpp
@@ -12,167 +12,6 @@ struct distribution_test_init
 static distribution_test_init const distrib_test_info;
 #endif
 
-template<typename P>
-void test_kronmult(prog_opts const &opts, P const tol_factor)
-{
-  discretization_manager disc(make_PDE<P>(opts));
-
-  auto &pde = disc.get_pde();
-
-  // assume uniform degree across dimensions
-  int const degree = disc.degree();
-
-  adapt::distributed_grid<P> const &grid = disc.get_grid();
-  elements::table const &table = grid.get_table();
-  element_subgrid const my_subgrid(0, table.size() - 1, 0, table.size() - 1);
-
-  // setup x vector
-  unsigned int seed{666};
-  std::mt19937 mersenne_engine(seed);
-  std::uniform_int_distribution<int> dist(-4, 4);
-  auto const gen = [&dist, &mersenne_engine]() {
-    return dist(mersenne_engine);
-  };
-  auto const elem_size  = fm::ipow(degree + 1, pde.num_dims());
-  fk::vector<P> const b = [&, gen]() {
-    fk::vector<P> output(disc.state_size());
-    std::generate(output.begin(), output.end(), gen);
-    return output;
-  }();
-
-  fk::vector<P> const gold = [&]() {
-    int64_t const system_size = disc.state_size();
-    fk::matrix<P> A(system_size, system_size);
-    fk::vector<P> x(b);
-    build_system_matrix<P>(
-        pde, [&](int t, int d)->fk::matrix<P>{ return disc.get_coeff_matrix(t, d); },
-        table, A, my_subgrid);
-    // AA = I - dt*A;
-    fm::scal(P{-1.} * pde.get_dt(), A);
-    for (int i = 0; i < A.nrows(); ++i)
-    {
-      A(i, i) += 1.0;
-    }
-    std::vector<int> ipiv(A.nrows());
-    fm::gesv(A, x, ipiv);
-    return x;
-  }();
-
-  // perform gmres with system matrix A
-  fk::vector<P> const gmres = [&pde, &table, &my_subgrid, &b, &disc,
-                               elem_size]() {
-    auto const system_size = elem_size * table.size();
-    fk::matrix<P> A(system_size, system_size);
-    fk::vector<P> x(b);
-    int const restart  = solvers::novalue;
-    int const max_iter = solvers::novalue;
-    P const tolerance  = solvers::notolerance;
-    build_system_matrix<P>(
-        pde, [&](int t, int d)->fk::matrix<P>{ return disc.get_coeff_matrix(t, d); },
-        table, A, my_subgrid);
-    // AA = I - dt*A;
-    fm::scal(P{-1.} * pde.get_dt(), A);
-    for (int i = 0; i < A.nrows(); ++i)
-    {
-      A(i, i) += 1.0;
-    }
-    solvers::simple_gmres(A, x, b, fk::matrix<P>(), restart, max_iter,
-                         tolerance);
-    return x;
-  }();
-
-  rmse_comparison(gold, gmres, tol_factor);
-
-  // perform bicgstab with system matrix A
-  fk::vector<P> const bicgstab = [&pde, &table, &my_subgrid, &b, &disc,
-                                  elem_size]() {
-    auto const system_size = elem_size * table.size();
-    fk::matrix<P> A(system_size, system_size);
-    fk::vector<P> x(b);
-    int const max_iter = solvers::novalue;
-    P const tolerance  = solvers::notolerance;
-    build_system_matrix<P>(
-        pde, [&](int t, int d)->fk::matrix<P>{ return disc.get_coeff_matrix(t, d); },
-        table, A, my_subgrid);
-    // AA = I - dt*A;
-    fm::scal(P{-1.} * pde.get_dt(), A);
-    for (int i = 0; i < A.nrows(); ++i)
-    {
-      A(i, i) += 1.0;
-    }
-    solvers::bicgstab(A, x, b, fk::matrix<P>(), max_iter,
-                      tolerance);
-    return x;
-  }();
-
-  rmse_comparison(gold, bicgstab, tol_factor);
-
-  asgard::kron_operators<P> &operator_matrices = disc.get_kronops();
-  operator_matrices.make(imex_flag::unspecified, pde, disc.get_cmatrices(), grid);
-
-  P const dt = pde.get_dt();
-
-  // perform matrix-free gmres
-  fk::vector<P> const matrix_free_gmres = [&operator_matrices, &b,
-                                           dt]() {
-    fk::vector<P> x(b);
-    int const restart  = solvers::novalue;
-    int const max_iter = solvers::novalue;
-    P const tolerance  = std::is_same_v<float, P> ? 1e-6 : 1e-12;
-    solvers::simple_gmres_euler(dt, imex_flag::unspecified, operator_matrices, x,
-                                b, restart, max_iter, tolerance);
-    return x;
-  }();
-
-  rmse_comparison(gold, matrix_free_gmres, tol_factor);
-
-  // perform matrix-free bicgstab
-  fk::vector<P> const matrix_free_bicgstab = [&operator_matrices, &b,
-                                              dt]() {
-    fk::vector<P> x(b);
-    int const max_iter = solvers::novalue;
-    P const tolerance  = std::is_same_v<float, P> ? 1e-6 : 1e-12;
-    solvers::bicgstab_euler(dt, imex_flag::unspecified, operator_matrices, x,
-                            b, max_iter, tolerance);
-    return x;
-  }();
-
-  rmse_comparison(gold, matrix_free_bicgstab, tol_factor);
-
-#ifdef ASGARD_USE_CUDA
-  // perform matrix-free gmres
-  fk::vector<P> const mf_gpu_gmres = [&operator_matrices, &gold, &b, dt]() {
-    fk::vector<P, mem_type::owner, resource::device> x_d =
-        b.clone_onto_device();
-    fk::vector<P, mem_type::owner, resource::device> b_d =
-        b.clone_onto_device();
-    int const restart  = solvers::novalue;
-    int const max_iter = solvers::novalue;
-    P const tolerance  = std::is_same_v<float, P> ? 1e-6 : 1e-12;
-    solvers::simple_gmres_euler(dt, imex_flag::unspecified, operator_matrices,
-                                x_d, b_d, restart, max_iter, tolerance);
-    return x_d.clone_onto_host();
-  }();
-
-  rmse_comparison(gold, mf_gpu_gmres, tol_factor);
-
-  // perform matrix-free bicgstab
-  fk::vector<P> const mf_gpu_bicgstab = [&operator_matrices, &gold, &b, dt]() {
-    fk::vector<P, mem_type::owner, resource::device> x_d =
-        b.clone_onto_device();
-    fk::vector<P, mem_type::owner, resource::device> b_d =
-        b.clone_onto_device();
-    int const max_iter = solvers::novalue;
-    P const tolerance  = std::is_same_v<float, P> ? 1e-6 : 1e-12;
-    solvers::bicgstab_euler(dt, imex_flag::unspecified, operator_matrices,
-                            x_d, b_d, max_iter, tolerance);
-    return x_d.clone_onto_host();
-  }();
-
-  rmse_comparison(gold, mf_gpu_bicgstab, tol_factor);
-#endif
-}
-
 TEMPLATE_TEST_CASE("simple GMRES", "[solver]", test_precs)
 {
   fk::matrix<TestType> const A_gold{
