--- testing/generated-inputs/element_table/1d_index_0_0.dat
@@ -1,4 +1,0 @@
-# Created from matlab git hash 78dc986b69a3e9573c0f04a335fe44be873f857e
-# name: foo
-# type: scalar
-1
