--- testing/generated-inputs/element_table/1d_index_2_1.dat
@@ -1,4 +1,0 @@
-# Created from matlab git hash cb08dc4c0d742d9e6843931db11a44da8739469d
-# name: foo
-# type: scalar
-4
