--- testing/generated-inputs/element_table/child_ids_2d_SG.dat
@@ -1,225 +1,0 @@
-# Created from matlab git hash 5029884412036be83b4f6a559a60e9e75dbeae9f
-# name: foo
-# type: matrix
-# rows: 220
-# columns: 1
-32
-33
-144
-34
-35
-145
-36
-37
-146
-38
-39
-147
-40
-41
-148
-42
-43
-149
-44
-45
-150
-46
-47
-151
-48
-49
-152
-50
-51
-153
-52
-53
-154
-54
-55
-155
-56
-57
-156
-58
-59
-157
-60
-61
-158
-62
-63
-159
-144
-145
-264
-392
-146
-147
-265
-393
-148
-149
-266
-394
-150
-151
-267
-395
-152
-153
-268
-396
-154
-155
-269
-397
-156
-157
-270
-398
-158
-159
-271
-399
-264
-265
-516
-644
-266
-267
-517
-645
-268
-269
-518
-646
-270
-271
-519
-647
-392
-393
-772
-900
-394
-395
-773
-901
-396
-397
-774
-902
-398
-399
-775
-903
-16
-17
-136
-18
-19
-137
-20
-21
-138
-22
-23
-139
-24
-25
-140
-26
-27
-141
-28
-29
-142
-30
-31
-143
-136
-137
-260
-388
-138
-139
-261
-389
-140
-141
-262
-390
-142
-143
-263
-391
-260
-261
-514
-642
-262
-263
-515
-643
-388
-389
-770
-898
-390
-391
-771
-899
-8
-9
-132
-10
-11
-133
-12
-13
-134
-14
-15
-135
-132
-133
-258
-386
-134
-135
-259
-387
-258
-259
-513
-641
-386
-387
-769
-897
-4
-5
-130
-6
-7
-131
-130
-131
-257
-385
-257
-512
-640
-385
-768
-896
-2
-3
-129
-129
-256
-384
-1
-128
