--- testing/generated-inputs/element_table/child_ids_3d_SG.dat
@@ -1,157 +1,0 @@
-# Created from matlab git hash 5029884412036be83b4f6a559a60e9e75dbeae9f
-# name: foo
-# type: matrix
-# rows: 152
-# columns: 1
-8
-9
-132
-16388
-10
-11
-133
-16389
-12
-13
-134
-16390
-14
-15
-135
-16391
-132
-133
-258
-386
-16514
-134
-135
-259
-387
-16515
-258
-259
-513
-641
-16641
-386
-387
-769
-897
-16769
-16388
-16389
-16514
-32770
-49154
-16390
-16391
-16515
-32771
-49155
-16514
-16515
-16641
-16769
-32897
-49281
-16641
-16896
-17024
-33024
-49408
-16769
-17152
-17280
-33152
-49536
-32770
-32771
-32897
-65537
-81921
-49154
-49155
-49281
-98305
-114689
-32897
-33024
-33152
-65664
-82048
-49281
-49408
-49536
-98432
-114816
-65537
-65664
-131072
-147456
-81921
-82048
-163840
-180224
-98305
-98432
-196608
-212992
-114689
-114816
-229376
-245760
-4
-5
-130
-16386
-6
-7
-131
-16387
-130
-131
-257
-385
-16513
-257
-512
-640
-16640
-385
-768
-896
-16768
-16386
-16387
-16513
-32769
-49153
-16513
-16640
-16768
-32896
-49280
-32769
-32896
-65536
-81920
-49153
-49280
-98304
-114688
-2
-3
-129
-16385
-129
-256
-384
-16512
-16385
-16512
-32768
-49152
-1
-128
-16384
