--- testing/generated-inputs/element_table/ids_1d_SG.dat
@@ -1,133 +1,0 @@
-# Created from matlab git hash 35e1f0babe3f66a54cae995de01de3ac542832e8
-# name: foo
-# type: matrix
-# rows: 128
-# columns: 1
-64
-65
-66
-67
-68
-69
-70
-71
-72
-73
-74
-75
-76
-77
-78
-79
-80
-81
-82
-83
-84
-85
-86
-87
-88
-89
-90
-91
-92
-93
-94
-95
-96
-97
-98
-99
-100
-101
-102
-103
-104
-105
-106
-107
-108
-109
-110
-111
-112
-113
-114
-115
-116
-117
-118
-119
-120
-121
-122
-123
-124
-125
-126
-127
-32
-33
-34
-35
-36
-37
-38
-39
-40
-41
-42
-43
-44
-45
-46
-47
-48
-49
-50
-51
-52
-53
-54
-55
-56
-57
-58
-59
-60
-61
-62
-63
-16
-17
-18
-19
-20
-21
-22
-23
-24
-25
-26
-27
-28
-29
-30
-31
-8
-9
-10
-11
-12
-13
-14
-15
-4
-5
-6
-7
-2
-3
-1
-0
