--- testing/generated-inputs/element_table/ids_2d_SG.dat
@@ -1,69 +1,0 @@
-# Created from matlab git hash 35e1f0babe3f66a54cae995de01de3ac542832e8
-# name: foo
-# type: matrix
-# rows: 64
-# columns: 1
-16
-17
-18
-19
-20
-21
-22
-23
-24
-25
-26
-27
-28
-29
-30
-31
-136
-137
-138
-139
-140
-141
-142
-143
-260
-261
-262
-263
-388
-389
-390
-391
-8
-9
-10
-11
-12
-13
-14
-15
-132
-133
-134
-135
-258
-259
-386
-387
-4
-5
-6
-7
-130
-131
-257
-385
-2
-3
-129
-256
-384
-1
-128
-0
