--- testing/generated-inputs/element_table/ids_3d_FG.dat
@@ -1,261 +1,0 @@
-# Created from matlab git hash 35e1f0babe3f66a54cae995de01de3ac542832e8
-# name: foo
-# type: matrix
-# rows: 256
-# columns: 1
-0
-1
-2
-3
-4
-5
-6
-7
-128
-129
-130
-131
-132
-133
-134
-135
-256
-384
-257
-385
-258
-259
-386
-387
-260
-261
-262
-263
-388
-389
-390
-391
-16384
-16385
-16386
-16387
-16388
-16389
-16390
-16391
-16512
-16513
-16514
-16515
-16516
-16517
-16518
-16519
-16640
-16768
-16641
-16769
-16642
-16643
-16770
-16771
-16644
-16645
-16646
-16647
-16772
-16773
-16774
-16775
-32768
-49152
-32769
-49153
-32770
-32771
-49154
-49155
-32772
-32773
-32774
-32775
-49156
-49157
-49158
-49159
-32896
-49280
-32897
-49281
-32898
-32899
-49282
-49283
-32900
-32901
-32902
-32903
-49284
-49285
-49286
-49287
-33024
-33152
-49408
-49536
-33025
-33153
-49409
-49537
-33026
-33027
-33154
-33155
-49410
-49411
-49538
-49539
-33028
-33029
-33030
-33031
-33156
-33157
-33158
-33159
-49412
-49413
-49414
-49415
-49540
-49541
-49542
-49543
-65536
-81920
-98304
-114688
-65537
-81921
-98305
-114689
-65538
-65539
-81922
-81923
-98306
-98307
-114690
-114691
-65540
-65541
-65542
-65543
-81924
-81925
-81926
-81927
-98308
-98309
-98310
-98311
-114692
-114693
-114694
-114695
-65664
-82048
-98432
-114816
-65665
-82049
-98433
-114817
-65666
-65667
-82050
-82051
-98434
-98435
-114818
-114819
-65668
-65669
-65670
-65671
-82052
-82053
-82054
-82055
-98436
-98437
-98438
-98439
-114820
-114821
-114822
-114823
-65792
-65920
-82176
-82304
-98560
-98688
-114944
-115072
-65793
-65921
-82177
-82305
-98561
-98689
-114945
-115073
-65794
-65795
-65922
-65923
-82178
-82179
-82306
-82307
-98562
-98563
-98690
-98691
-114946
-114947
-115074
-115075
-65796
-65797
-65798
-65799
-65924
-65925
-65926
-65927
-82180
-82181
-82182
-82183
-82308
-82309
-82310
-82311
-98564
-98565
-98566
-98567
-98692
-98693
-98694
-98695
-114948
-114949
-114950
-114951
-115076
-115077
-115078
-115079
