--- testing/generated-inputs/element_table/ids_3d_SG.dat
@@ -1,39 +1,0 @@
-# Created from matlab git hash 35e1f0babe3f66a54cae995de01de3ac542832e8
-# name: foo
-# type: matrix
-# rows: 34
-# columns: 1
-4
-5
-6
-7
-130
-131
-257
-385
-16386
-16387
-16513
-16640
-16768
-32769
-49153
-32896
-49280
-65536
-81920
-98304
-114688
-2
-3
-129
-256
-384
-16385
-16512
-32768
-49152
-1
-128
-16384
-0
