--- testing/generated-inputs/element_table/table_1d_FG.dat
@@ -1,133 +1,0 @@
-# Created from matlab git hash 35e1f0babe3f66a54cae995de01de3ac542832e8
-# name: foo
-# type: matrix
-# rows: 128
-# columns: 2
-0 0
-1 0
-2 0
-2 1
-3 0
-3 1
-3 2
-3 3
-4 0
-4 1
-4 2
-4 3
-4 4
-4 5
-4 6
-4 7
-5 0
-5 1
-5 2
-5 3
-5 4
-5 5
-5 6
-5 7
-5 8
-5 9
-5 10
-5 11
-5 12
-5 13
-5 14
-5 15
-6 0
-6 1
-6 2
-6 3
-6 4
-6 5
-6 6
-6 7
-6 8
-6 9
-6 10
-6 11
-6 12
-6 13
-6 14
-6 15
-6 16
-6 17
-6 18
-6 19
-6 20
-6 21
-6 22
-6 23
-6 24
-6 25
-6 26
-6 27
-6 28
-6 29
-6 30
-6 31
-7 0
-7 1
-7 2
-7 3
-7 4
-7 5
-7 6
-7 7
-7 8
-7 9
-7 10
-7 11
-7 12
-7 13
-7 14
-7 15
-7 16
-7 17
-7 18
-7 19
-7 20
-7 21
-7 22
-7 23
-7 24
-7 25
-7 26
-7 27
-7 28
-7 29
-7 30
-7 31
-7 32
-7 33
-7 34
-7 35
-7 36
-7 37
-7 38
-7 39
-7 40
-7 41
-7 42
-7 43
-7 44
-7 45
-7 46
-7 47
-7 48
-7 49
-7 50
-7 51
-7 52
-7 53
-7 54
-7 55
-7 56
-7 57
-7 58
-7 59
-7 60
-7 61
-7 62
-7 63
