--- testing/generated-inputs/element_table/table_2d_FG.dat
@@ -1,133 +1,0 @@
-# Created from matlab git hash 35e1f0babe3f66a54cae995de01de3ac542832e8
-# name: foo
-# type: matrix
-# rows: 128
-# columns: 4
-0 0 0 0
-1 0 0 0
-2 0 0 0
-2 0 1 0
-3 0 0 0
-3 0 1 0
-3 0 2 0
-3 0 3 0
-4 0 0 0
-4 0 1 0
-4 0 2 0
-4 0 3 0
-4 0 4 0
-4 0 5 0
-4 0 6 0
-4 0 7 0
-5 0 0 0
-5 0 1 0
-5 0 2 0
-5 0 3 0
-5 0 4 0
-5 0 5 0
-5 0 6 0
-5 0 7 0
-5 0 8 0
-5 0 9 0
-5 0 10 0
-5 0 11 0
-5 0 12 0
-5 0 13 0
-5 0 14 0
-5 0 15 0
-0 1 0 0
-1 1 0 0
-2 1 0 0
-2 1 1 0
-3 1 0 0
-3 1 1 0
-3 1 2 0
-3 1 3 0
-4 1 0 0
-4 1 1 0
-4 1 2 0
-4 1 3 0
-4 1 4 0
-4 1 5 0
-4 1 6 0
-4 1 7 0
-5 1 0 0
-5 1 1 0
-5 1 2 0
-5 1 3 0
-5 1 4 0
-5 1 5 0
-5 1 6 0
-5 1 7 0
-5 1 8 0
-5 1 9 0
-5 1 10 0
-5 1 11 0
-5 1 12 0
-5 1 13 0
-5 1 14 0
-5 1 15 0
-0 2 0 0
-0 2 0 1
-1 2 0 0
-1 2 0 1
-2 2 0 0
-2 2 1 0
-2 2 0 1
-2 2 1 1
-3 2 0 0
-3 2 1 0
-3 2 2 0
-3 2 3 0
-3 2 0 1
-3 2 1 1
-3 2 2 1
-3 2 3 1
-4 2 0 0
-4 2 1 0
-4 2 2 0
-4 2 3 0
-4 2 4 0
-4 2 5 0
-4 2 6 0
-4 2 7 0
-4 2 0 1
-4 2 1 1
-4 2 2 1
-4 2 3 1
-4 2 4 1
-4 2 5 1
-4 2 6 1
-4 2 7 1
-5 2 0 0
-5 2 1 0
-5 2 2 0
-5 2 3 0
-5 2 4 0
-5 2 5 0
-5 2 6 0
-5 2 7 0
-5 2 8 0
-5 2 9 0
-5 2 10 0
-5 2 11 0
-5 2 12 0
-5 2 13 0
-5 2 14 0
-5 2 15 0
-5 2 0 1
-5 2 1 1
-5 2 2 1
-5 2 3 1
-5 2 4 1
-5 2 5 1
-5 2 6 1
-5 2 7 1
-5 2 8 1
-5 2 9 1
-5 2 10 1
-5 2 11 1
-5 2 12 1
-5 2 13 1
-5 2 14 1
-5 2 15 1
