--- testing/generated-inputs/element_table/table_2d_SG.dat
@@ -1,69 +1,0 @@
-# Created from matlab git hash 35e1f0babe3f66a54cae995de01de3ac542832e8
-# name: foo
-# type: matrix
-# rows: 64
-# columns: 4
-5 0 0 0
-5 0 1 0
-5 0 2 0
-5 0 3 0
-5 0 4 0
-5 0 5 0
-5 0 6 0
-5 0 7 0
-5 0 8 0
-5 0 9 0
-5 0 10 0
-5 0 11 0
-5 0 12 0
-5 0 13 0
-5 0 14 0
-5 0 15 0
-4 1 0 0
-4 1 1 0
-4 1 2 0
-4 1 3 0
-4 1 4 0
-4 1 5 0
-4 1 6 0
-4 1 7 0
-3 2 0 0
-3 2 1 0
-3 2 2 0
-3 2 3 0
-3 2 0 1
-3 2 1 1
-3 2 2 1
-3 2 3 1
-4 0 0 0
-4 0 1 0
-4 0 2 0
-4 0 3 0
-4 0 4 0
-4 0 5 0
-4 0 6 0
-4 0 7 0
-3 1 0 0
-3 1 1 0
-3 1 2 0
-3 1 3 0
-2 2 0 0
-2 2 1 0
-2 2 0 1
-2 2 1 1
-3 0 0 0
-3 0 1 0
-3 0 2 0
-3 0 3 0
-2 1 0 0
-2 1 1 0
-1 2 0 0
-1 2 0 1
-2 0 0 0
-2 0 1 0
-1 1 0 0
-0 2 0 0
-0 2 0 1
-1 0 0 0
-0 1 0 0
-0 0 0 0
