--- testing/generated-inputs/element_table/table_3d_FG.dat
@@ -1,261 +1,0 @@
-# Created from matlab git hash 35e1f0babe3f66a54cae995de01de3ac542832e8
-# name: foo
-# type: matrix
-# rows: 256
-# columns: 6
-0 0 0 0 0 0
-1 0 0 0 0 0
-2 0 0 0 0 0
-2 0 0 1 0 0
-3 0 0 0 0 0
-3 0 0 1 0 0
-3 0 0 2 0 0
-3 0 0 3 0 0
-0 1 0 0 0 0
-1 1 0 0 0 0
-2 1 0 0 0 0
-2 1 0 1 0 0
-3 1 0 0 0 0
-3 1 0 1 0 0
-3 1 0 2 0 0
-3 1 0 3 0 0
-0 2 0 0 0 0
-0 2 0 0 1 0
-1 2 0 0 0 0
-1 2 0 0 1 0
-2 2 0 0 0 0
-2 2 0 1 0 0
-2 2 0 0 1 0
-2 2 0 1 1 0
-3 2 0 0 0 0
-3 2 0 1 0 0
-3 2 0 2 0 0
-3 2 0 3 0 0
-3 2 0 0 1 0
-3 2 0 1 1 0
-3 2 0 2 1 0
-3 2 0 3 1 0
-0 0 1 0 0 0
-1 0 1 0 0 0
-2 0 1 0 0 0
-2 0 1 1 0 0
-3 0 1 0 0 0
-3 0 1 1 0 0
-3 0 1 2 0 0
-3 0 1 3 0 0
-0 1 1 0 0 0
-1 1 1 0 0 0
-2 1 1 0 0 0
-2 1 1 1 0 0
-3 1 1 0 0 0
-3 1 1 1 0 0
-3 1 1 2 0 0
-3 1 1 3 0 0
-0 2 1 0 0 0
-0 2 1 0 1 0
-1 2 1 0 0 0
-1 2 1 0 1 0
-2 2 1 0 0 0
-2 2 1 1 0 0
-2 2 1 0 1 0
-2 2 1 1 1 0
-3 2 1 0 0 0
-3 2 1 1 0 0
-3 2 1 2 0 0
-3 2 1 3 0 0
-3 2 1 0 1 0
-3 2 1 1 1 0
-3 2 1 2 1 0
-3 2 1 3 1 0
-0 0 2 0 0 0
-0 0 2 0 0 1
-1 0 2 0 0 0
-1 0 2 0 0 1
-2 0 2 0 0 0
-2 0 2 1 0 0
-2 0 2 0 0 1
-2 0 2 1 0 1
-3 0 2 0 0 0
-3 0 2 1 0 0
-3 0 2 2 0 0
-3 0 2 3 0 0
-3 0 2 0 0 1
-3 0 2 1 0 1
-3 0 2 2 0 1
-3 0 2 3 0 1
-0 1 2 0 0 0
-0 1 2 0 0 1
-1 1 2 0 0 0
-1 1 2 0 0 1
-2 1 2 0 0 0
-2 1 2 1 0 0
-2 1 2 0 0 1
-2 1 2 1 0 1
-3 1 2 0 0 0
-3 1 2 1 0 0
-3 1 2 2 0 0
-3 1 2 3 0 0
-3 1 2 0 0 1
-3 1 2 1 0 1
-3 1 2 2 0 1
-3 1 2 3 0 1
-0 2 2 0 0 0
-0 2 2 0 1 0
-0 2 2 0 0 1
-0 2 2 0 1 1
-1 2 2 0 0 0
-1 2 2 0 1 0
-1 2 2 0 0 1
-1 2 2 0 1 1
-2 2 2 0 0 0
-2 2 2 1 0 0
-2 2 2 0 1 0
-2 2 2 1 1 0
-2 2 2 0 0 1
-2 2 2 1 0 1
-2 2 2 0 1 1
-2 2 2 1 1 1
-3 2 2 0 0 0
-3 2 2 1 0 0
-3 2 2 2 0 0
-3 2 2 3 0 0
-3 2 2 0 1 0
-3 2 2 1 1 0
-3 2 2 2 1 0
-3 2 2 3 1 0
-3 2 2 0 0 1
-3 2 2 1 0 1
-3 2 2 2 0 1
-3 2 2 3 0 1
-3 2 2 0 1 1
-3 2 2 1 1 1
-3 2 2 2 1 1
-3 2 2 3 1 1
-0 0 3 0 0 0
-0 0 3 0 0 1
-0 0 3 0 0 2
-0 0 3 0 0 3
-1 0 3 0 0 0
-1 0 3 0 0 1
-1 0 3 0 0 2
-1 0 3 0 0 3
-2 0 3 0 0 0
-2 0 3 1 0 0
-2 0 3 0 0 1
-2 0 3 1 0 1
-2 0 3 0 0 2
-2 0 3 1 0 2
-2 0 3 0 0 3
-2 0 3 1 0 3
-3 0 3 0 0 0
-3 0 3 1 0 0
-3 0 3 2 0 0
-3 0 3 3 0 0
-3 0 3 0 0 1
-3 0 3 1 0 1
-3 0 3 2 0 1
-3 0 3 3 0 1
-3 0 3 0 0 2
-3 0 3 1 0 2
-3 0 3 2 0 2
-3 0 3 3 0 2
-3 0 3 0 0 3
-3 0 3 1 0 3
-3 0 3 2 0 3
-3 0 3 3 0 3
-0 1 3 0 0 0
-0 1 3 0 0 1
-0 1 3 0 0 2
-0 1 3 0 0 3
-1 1 3 0 0 0
-1 1 3 0 0 1
-1 1 3 0 0 2
-1 1 3 0 0 3
-2 1 3 0 0 0
-2 1 3 1 0 0
-2 1 3 0 0 1
-2 1 3 1 0 1
-2 1 3 0 0 2
-2 1 3 1 0 2
-2 1 3 0 0 3
-2 1 3 1 0 3
-3 1 3 0 0 0
-3 1 3 1 0 0
-3 1 3 2 0 0
-3 1 3 3 0 0
-3 1 3 0 0 1
-3 1 3 1 0 1
-3 1 3 2 0 1
-3 1 3 3 0 1
-3 1 3 0 0 2
-3 1 3 1 0 2
-3 1 3 2 0 2
-3 1 3 3 0 2
-3 1 3 0 0 3
-3 1 3 1 0 3
-3 1 3 2 0 3
-3 1 3 3 0 3
-0 2 3 0 0 0
-0 2 3 0 1 0
-0 2 3 0 0 1
-0 2 3 0 1 1
-0 2 3 0 0 2
-0 2 3 0 1 2
-0 2 3 0 0 3
-0 2 3 0 1 3
-1 2 3 0 0 0
-1 2 3 0 1 0
-1 2 3 0 0 1
-1 2 3 0 1 1
-1 2 3 0 0 2
-1 2 3 0 1 2
-1 2 3 0 0 3
-1 2 3 0 1 3
-2 2 3 0 0 0
-2 2 3 1 0 0
-2 2 3 0 1 0
-2 2 3 1 1 0
-2 2 3 0 0 1
-2 2 3 1 0 1
-2 2 3 0 1 1
-2 2 3 1 1 1
-2 2 3 0 0 2
-2 2 3 1 0 2
-2 2 3 0 1 2
-2 2 3 1 1 2
-2 2 3 0 0 3
-2 2 3 1 0 3
-2 2 3 0 1 3
-2 2 3 1 1 3
-3 2 3 0 0 0
-3 2 3 1 0 0
-3 2 3 2 0 0
-3 2 3 3 0 0
-3 2 3 0 1 0
-3 2 3 1 1 0
-3 2 3 2 1 0
-3 2 3 3 1 0
-3 2 3 0 0 1
-3 2 3 1 0 1
-3 2 3 2 0 1
-3 2 3 3 0 1
-3 2 3 0 1 1
-3 2 3 1 1 1
-3 2 3 2 1 1
-3 2 3 3 1 1
-3 2 3 0 0 2
-3 2 3 1 0 2
-3 2 3 2 0 2
-3 2 3 3 0 2
-3 2 3 0 1 2
-3 2 3 1 1 2
-3 2 3 2 1 2
-3 2 3 3 1 2
-3 2 3 0 0 3
-3 2 3 1 0 3
-3 2 3 2 0 3
-3 2 3 3 0 3
-3 2 3 0 1 3
-3 2 3 1 1 3
-3 2 3 2 1 3
-3 2 3 3 1 3
