--- testing/generated-inputs/element_table/table_3d_SG.dat
@@ -1,39 +1,0 @@
-# Created from matlab git hash 35e1f0babe3f66a54cae995de01de3ac542832e8
-# name: foo
-# type: matrix
-# rows: 34
-# columns: 6
-3 0 0 0 0 0
-3 0 0 1 0 0
-3 0 0 2 0 0
-3 0 0 3 0 0
-2 1 0 0 0 0
-2 1 0 1 0 0
-1 2 0 0 0 0
-1 2 0 0 1 0
-2 0 1 0 0 0
-2 0 1 1 0 0
-1 1 1 0 0 0
-0 2 1 0 0 0
-0 2 1 0 1 0
-1 0 2 0 0 0
-1 0 2 0 0 1
-0 1 2 0 0 0
-0 1 2 0 0 1
-0 0 3 0 0 0
-0 0 3 0 0 1
-0 0 3 0 0 2
-0 0 3 0 0 3
-2 0 0 0 0 0
-2 0 0 1 0 0
-1 1 0 0 0 0
-0 2 0 0 0 0
-0 2 0 0 1 0
-1 0 1 0 0 0
-0 1 1 0 0 0
-0 0 2 0 0 0
-0 0 2 0 0 1
-1 0 0 0 0 0
-0 1 0 0 0 0
-0 0 1 0 0 0
-0 0 0 0 0 0
